requirements.txt
accelerate==0.32.0
albumentations==1.4.18
decord==0.6.0
diffusers==0.19.0
einops==0.8.0
huggingface-hub==0.21.0
imageio==2.34.2
matplotlib==3.7.5
natsort
numpy==1.24.4
opencv-python==4.*
pillow==10.4.0
requests==2.32.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
tensorboard==2.14.0
tokenizers==0.15.2
torch==2.1.1
torch-dct==0.1.6
torchvision==0.16.1
tqdm==4.66.4
transformers==4.35.2
triton==2.1.0
wheel==0.43.0
xformers==0.0.23